requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
pillow==10.3.0
qrcode[pil]==7.4.2
python-barcode==0.14.0  